Remove aic(4) driver
aic was marked to be gone in 12 a while ago. Go ahead and remove it.
nycbug's dmesg database shows this was last seen in 6 and one more
time in 4.x. It never was popular, and what popularity it had was over
before the nycbug databse got going in 2004.

Relnotes: yes


git-svn-id: https://svn.freebsd.org/base/head@339566 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f