Remove bt(4) driver
The buslogic scsi driver has been tagged as gone in 12 for some time
now. Remove it. The nycbug dmesg database shows only one sighting in 6
for this driver. It was very popular in the early days of the project,
but that popularity seems to have died by 2004 when the nycbug
database started up.

Relnotes: yes


git-svn-id: https://svn.freebsd.org/base/head@339568 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f