cxgbe(4): Allow "pass" filters to distribute matching traffic using a subset of a VI's RSS indirection table.
This makes it possible to make groups out of rx queues and steer
different kinds of traffic to different groups.  For example, an
interface with 8 rx queues could have all non-TCP traffic delivered to
queues 0-3 and all TCP traffic to queues 4-7.

Note that it is already possible for filters to steer traffic to a
particular queue or to distribute it using the full indirection table
(much like normal rx does).

Sponsored by:	Chelsio Communications


git-svn-id: https://svn.freebsd.org/base/head@339717 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f