MFC r339849: lualoader: Fix try_include error handling
The previous iteration of try_include attempted to be 'friendly' and error()
out if we hit an error that wasn't ENOENT. This was semi-OK, but fragile as
it relied on pattern matching the error message.

Move the responsibility for handling failure to the caller. Following
a common lua pattern, we'll return the return value of the underlying
require() on success, or false and an error message.

Approved by:	re (gjb)


git-svn-id: https://svn.freebsd.org/base/stable/12@340012 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f