MFC r340236:
Fix rxcsum issue introduced in r338838

r338838 attempted to fix issues with rxcsum and rxcsum6.
However, the rxcsum bits were set as though if_setcapenablebit() was
being called, not if_togglecapenable() which is in use. As a result,
it was not possible to disable rxcsum when rxcsum6 was supported.

PR:		233004
Reported by:	lev
Reviewed by:	lev
Approved by:	re (kib)
Sponsored by:	Limelight Networks
Differential Revision:	https://reviews.freebsd.org/D17881


git-svn-id: https://svn.freebsd.org/base/stable/12@340365 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f