MFC r339470:
pf synproxy will do the 3WHS on behalf of the target machine, and once
the 3WHS is completed, establish the backend connection. The trigger
for "3WHS completed" is the reception of the first ACK. However, we
should not proceed if that ACK also has RST or FIN set.

PR:		197484
Obtained from:	OpenBSD


git-svn-id: https://svn.freebsd.org/base/stable/11@340559 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f