MFC r340068:
pfsync: Handle syncdev going away

If the syncdev is removed we no longer need to clean up the multicast
entry we've got set up for that device.

Pass the ifnet detach event through pf to pfsync, and remove our
multicast handle, and mark us as no longer having a syncdev.

Note that this callback is always installed, even if the pfsync
interface is disabled (and thus it's not a per-vnet callback pointer).

Sponsored by:	Orange Business Services


git-svn-id: https://svn.freebsd.org/base/stable/12@340573 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f