MFC: r339999 Fix NFS client vnode locking to avoid a crash during forced dismount.
A crash was reported where the crash occurred in nfs_advlock() when the
NFS_ISV4(vp) macro was being executed. This was caused by the vnode
being VI_DOOMED due to a forced dismount in progress.
This patch fixes the problem by locking the vnode before executing the
NFS_ISV4() macro.

PR:		232673


git-svn-id: https://svn.freebsd.org/base/stable/12@340590 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f