bectl(8) tests: attempt to load the ZFS module
Observed in a CI test image, bectl_create test will run and be marked as
skipped because the module is not loaded. The first zpool invocation will
automagically load the module, but bectl_create is still skipped. Subsequent
tests all pass as expected because the module is now loaded and everything
is OK.

MFC after:	3 days


git-svn-id: https://svn.freebsd.org/base/head@340636 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f